requirements.txt
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
qrcode>=8.0
drf-spectacular>=0.27.2
django-environ>=0.11.2
django-cors-headers>=4.6.0